+++ src/main/java/at/hannibal2/skyhanni/config/features/GUIConfig.java
@@ -54,6 +54,12 @@ public class GUIConfig {
     public boolean realTime = false;
 
     @Expose
+    @ConfigOption(name = "Real Time 12h format", desc = "Display the current computer time in 12hr format rather than 24h format.")
+    @ConfigEditorBoolean
+    @FeatureToggle
+    public boolean realTimeFormatToggle = false;
+
+    @Expose
     public Position realTimePosition = new Position(10, 10, false, true);
 
     @Expose