+++ CHANGELOG.md
@@ -175,6 +175,7 @@
 + Improved performance of item ability cooldown and farming fortune display data. - Cad
     + The item flickering for ability cooldown background should be less/gone completely.
 + Added the ability to bypass not clickable items when holding the control key. - CalMWolfs
++ Added toggle for 12hr/24hr in real time hud - Thunderblade73
 
 ### Fixes